requirements.txt
cleverhans==4.0.0
matplotlib==3.5.2
numpy==1.21.5
pandas==1.4.4
Pillow==9.2.0
pytorch_metric_learning==2.0.1
PyYAML==6.0
scikit_learn==1.0.2
seaborn==0.11.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
umap_learn==0.5.3
